Milvus-compatible C++ gRPC server and client
Add a Milvus-compatible serving layer on top of PipeANN:
- gRPC server (src/server) with collection/doc stores, field codec,
  and a search worker pool, exposing the Milvus proto API
- Python milvus/server clients replacing the old qdrant/langchain/
  collection client surface
- Milvus vs PipeANN benchmarks and gRPC e2e + quickstart tests
- Simplified search_disk_index, DSL compiler for filter selectors,
  and assorted index/build cleanups

Squash-merge of dev-milvus-compatible (23 commits).

+++ src/search/beam_search.cpp
@@ -20,7 +20,7 @@
 
 namespace pipeann {
   template<typename T, typename TagT>
-  void SSDIndex<T, TagT>::do_beam_search(const T *query1, uint32_t mem_L, uint32_t l_search, const uint64_t beam_width,
+  void SSDIndex<T, TagT>::do_beam_search(const T *query1, uint32_t mem_L, uint64_t l_search, const uint64_t beam_width,
                                          std::vector<Neighbor> &expanded_nodes_info, QueryStats *stats,
                                          InsertContext *insert_ctx) {
     auto diskSearchBegin = std::chrono::high_resolution_clock::now();
@@ -225,7 +225,7 @@ namespace pipeann {
                                         const uint64_t beam_width, QueryStats *stats) {
     std::shared_lock lk(merge_lock);
     std::vector<Neighbor> expanded_nodes_info;
-    this->do_beam_search(query, mem_L, (uint32_t) l_search, beam_width, expanded_nodes_info, stats, nullptr);
+    this->do_beam_search(query, mem_L, l_search, beam_width, expanded_nodes_info, stats, nullptr);
 
     return copy_top_k(expanded_nodes_info, k_search, res_tags, distances);
   }
